Struct i_slint_backend_testing::ElementHandle

pub struct ElementHandle(/* private fields */);

ElementHandle`` wraps an existing element in a Slint UI. An ElementHandle does not keep the corresponding element in the UI alive. Use [Self::is_valid()`] to verify that it is still alive.

Obtain instances of ElementHandle by querying your application through Self::find_by_accessible_label().

Implementations

impl ElementHandle

pub fn is_valid(&self) -> bool

Returns true if the element still exists in the in UI and is valid to access; false otherwise.

pub fn find_by_accessible_label( component: &impl ComponentHandle, label: &str ) -> impl Iterator<Item = Self>

This function searches through the entire tree of elements of component, looks for elements that have a accessible-label property with the provided value label, and returns an iterator over the found elements.

pub fn invoke_accessible_default_action(&self)

Invokes the default accessible action on the element. For example a MyButton element might declare an accessible default action that simulates a click, as in the following example:

component MyButton {
    // ...
    callback clicked();
    in property <string> text;

    TouchArea {
        clicked => { root.clicked() }
    }
    accessible-role: button;
    accessible-label: self.text;
    accessible-action-default => { self.clicked(); }
}

pub fn accessible_value(&self) -> Option<SharedString>

Returns the value of the element’s accessible-value property, if present.

pub fn set_accessible_value(&self, value: impl Into<SharedString>)

Sets the value of the element’s accessible-value property. Note that you can only set this property if it is declared in your Slint code.

pub fn accessible_value_maximum(&self) -> Option<f32>

Returns the value of the element’s accessible-value-maximum property, if present.

pub fn accessible_value_minimum(&self) -> Option<f32>

Returns the value of the element’s accessible-value-minimum property, if present.

pub fn accessible_value_step(&self) -> Option<f32>

Returns the value of the element’s accessible-value-step property, if present.

pub fn accessible_label(&self) -> Option<SharedString>

Returns the value of the accessible-label property, if present.

pub fn accessible_description(&self) -> Option<SharedString>

Returns the value of the accessible-description property, if present

pub fn accessible_checked(&self) -> Option<bool>

Returns the value of the accessible-checked property, if present

pub fn accessible_checkable(&self) -> Option<bool>

Returns the value of the accessible-checkable property, if present

pub fn size(&self) -> LogicalSize

Returns the size of the element in logical pixels. This corresponds to the value of the width and height properties in Slint code. Returns a zero size if the element is not valid.

pub fn absolute_position(&self) -> LogicalPosition

Returns the position of the element within the entire window. This corresponds to the value of the absolute-position property in Slint code. Returns a zero position if the element is not valid.