i_slint_backend_testing

Enum AccessibleRole

#[non_exhaustive]

#[repr(u32)]
pub enum AccessibleRole {

    None = 0,
    Button = 1,
    Checkbox = 2,
    Combobox = 3,
    Groupbox = 4,
    Image = 5,
    List = 6,
    Slider = 7,
    Spinbox = 8,
    Tab = 9,
    TabList = 10,
    TabPanel = 11,
    Text = 12,
    Table = 13,
    Tree = 14,
    ProgressIndicator = 15,
    TextInput = 16,
    Switch = 17,
    ListItem = 18,

}

This enum represents the different values for the accessible-role property, used to describe the role of an element in the context of assistive technology such as screen readers.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

None = 0

The element isnโt accessible.

Button = 1

The element is a Button or behaves like one.

Checkbox = 2

The element is a CheckBox or behaves like one.

Combobox = 3

The element is a ComboBox or behaves like one.

Groupbox = 4

The element is a GroupBox or behaves like one.

Image = 5

The element is an Image or behaves like one. This is automatically applied to Image elements.

List = 6

The element is a ListView or behaves like one.

Slider = 7

The element is a Slider or behaves like one.

Spinbox = 8

The element is a SpinBox or behaves like one.

Tab = 9

The element is a Tab or behaves like one.

TabList = 10

The element is similar to the tab bar in a TabWidget.

TabPanel = 11

The element is a container for tab content.

Text = 12

The role for a Text element. This is automatically applied to Text elements.

Table = 13

The role for a TableView or behaves like one.

Tree = 14

The role for a TreeView or behaves like one. (Not provided yet)

ProgressIndicator = 15

The element is a ProgressIndicator or behaves like one.

TextInput = 16

The role for widget with editable text such as a LineEdit or a TextEdit. This is automatically applied to TextInput elements.

Switch = 17

The element is a Switch or behaves like one.

ListItem = 18

The element is an item in a ListView.